more work done
After much trouble with Inflating and such, I decided to move on and
not leave most of the work undone.  Decided to add an imagebutton that
when clicked opens the flixster webpage.  I decided on this because
when using the API it mentioned that using the data requires
representation of the API.  Thought it would be a good way to wrap it
up.