lots of research, less work
Read several articles in the android dev api, watched the lynda videos,
looked over some 3rd party tutorials on intents and passing data.  In
the end, the hardest part was ironing out a few bugs.  The last bug was
kind of a lucky guess.. I had to cast my getExtras to (String), and
then it worked.  Would have preferred the first way I set it up (a
stringArray), but this is acceptable for now.  really like what this
has added to my old app.