Wowie zowie!
Snap!  Everything works!  A few hours ago I wasn't sure if I'd turn in
a broken app or not.  The TutsPlus link was the clinch at the end.  I
had to make the 'dynamic choice'  to ignore the explicit intent to
activate the second activity.  Slapped in some populating code and
figured out a minor bug, and poof!  All done.