Super commit!
Ok I kinda forgot about commiting..  But it took me most of today to
get IntentService working properly anyways.  Had a helluva time with it
not finding my service properly.  After changing the service name in
the manifest a bunch, and many other things, one last change in the
manifest to the the path made it work suddenly!
After that I spent a rediculous amount of time trying to change the
look my my app...without result.  I was searching for a way to change
the theme of my second fragment when it is displayed with the
MainActivity.. but whatever I set the MainActivity's theme to overrides
it.  So after much XML guess work and not much help from Google I have
met the end of that road.  Perhaps I can get some insight from you
later.