Zee Grande Finale!~
Ok so.. Content providers, not so bad.  Near the end is when they
really started to come together for me.  Which is also why I decided
that I could not implement query into my app this week.  On the good
side I do have update working properly, and query -is- setup to work as
well.  It's just that under time constraints it wouldn't be feasable to
try to change my data implementation (from receiving pure json to a
cursor) **(have to get the cursor return down as well!)..  anyways,
good week overall!