Gee whiz.
Thats alot of stuff commiting..  Anyways, this was quite a long day
battling with action bars.  But through some means I was able to put
everything I wanted into place.  I had some maddening errors for a
while, so I basically rigged it to work.  "Removed" the spinner and
webIntent button, put them in the overflow of the action bar.  Spinner
is actually invisible and clicking 'History' on the action bar triggers
the click event for the spinner.  Nice way to sneak around that deal I
suppose.  Changed the web intent button to be on the bar just because..
why not??  Only took a brief minute.  Onto widgets!